requirements.txt
streamlit
streamlit-authenticator
pandas
python-dotenv
sqlalchemy
bcrypt
pytest
